Refactor JSON RPC variants
