+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/nethermind/trace.ex
